fixed the sync issue by not syncing

+++ src/main/java/net/dries007/tfc/common/blocks/fruit_tree/FruitTreeBranchBlock.java
@@ -76,7 +76,7 @@ public class FruitTreeBranchBlock extends SixWayBlock implements IForgeBlockProp
     @SuppressWarnings("deprecation")
     public void tick(BlockState state, ServerWorld worldIn, BlockPos pos, Random rand)
     {
-        if (!state.canSurvive(worldIn, pos))
+        if (!state.canSurvive(worldIn, pos) && !worldIn.isClientSide())
         {
             worldIn.destroyBlock(pos, true);
         }
